Move icp_get_blocks to test utils crate
